GroupServiceTest TaskServiceTest

+++ src/main/java/ee/taltech/iti0302/okapi/backend/services/TaskService.java
@@ -28,16 +28,16 @@ public class TaskService {
         return LocalDateTime.now();
     }
 
-    public List<TaskDTO> getAllTasks(int page) {
-        Sort sort = Sort.by("status").descending();
-        Pageable pageRequest = PageRequest.of(page, 10, sort);
-        Page<Task> task = taskRepository.findAll(pageRequest);
-        return task.getContent().stream()
-                .map(TaskMapper.INSTANCE::toDTO)
-                .toList();
-        log.info(getCurrentTime() + ": " + "Retrieved {} tasks", taskDTOs.size());
-        return taskDTOs;
-    }
+//    public List<TaskDTO> getAllTasks(int page) {
+//        Sort sort = Sort.by("status").descending();
+//        Pageable pageRequest = PageRequest.of(page, 10, sort);
+//        Page<Task> task = taskRepository.findAll(pageRequest);
+//        return task.getContent().stream()
+//                .map(TaskMapper.INSTANCE::toDTO)
+//                .toList();
+//        log.info(getCurrentTime() + ": " + "Retrieved {} tasks", taskDTOs.size());
+//        return taskDTOs;
+//    }
 
 
     public TaskDTO getTaskById(long id) {
@@ -50,15 +50,15 @@ public class TaskService {
     public TaskDTO createTask(TaskDTO dto) {
         log.info(getCurrentTime() + ": " + "Creating task: {}", dto.getTitle());
         Task task = taskRepository.save(TaskMapper.INSTANCE.toEntity(dto));
-        dto.setId(task.getId());
         if (task.getId() != null) {
+            dto.setId(task.getId());
             log.info(getCurrentTime() + ": " + "Task created successfully. Task ID: {}", task.getId());
             return dto;
-        } else {
-            log.warn(getCurrentTime() + ": " + "Task creation failed");
-            return null;
         }
+        log.warn(getCurrentTime() + ": " + "Task creation failed");
+        return null;
     }
+
 
     public TaskDTO updateTask(TaskDTO dto) {
         log.info(getCurrentTime() + ": " + "Updating task with ID: {}", dto.getId());
@@ -78,8 +78,12 @@ public class TaskService {
     }
 
     public void deleteTask(long id) {
-        taskRepository.deleteById(id);
-        log.info(getCurrentTime() + ": " + "Task deleted successfully. Task ID: {}", id);
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+            log.info(getCurrentTime() + ": " + "Task deleted successfully. Task ID: {}", id);
+        } else {
+            log.warn(getCurrentTime() + ": " + "Task deletion failed. Task not found with ID: {}", id);
+        }
     }
 
     public List<TaskDTO> findByTitle(int page, String title) {
